Refactor a couple functions (training time) to the constituency.utils package
